reccomendation: make the nonce documentaion more clear that there are two differnt counter, and maybe empty hashes for 7.1 (handshake key generation)

+++ key_schedule.h
@@ -6,7 +6,7 @@
     void process_client_handshake_traffic_secret(unsigned char* out, unsigned char* shared_secret, unsigned char* transcipt_hash);
     void process_base_derived_secret(unsigned char* out);
     void process_handshake_secret(unsigned char* out, unsigned char* shared_secret, unsigned char* derived_secret);
-    void generate_nonce(unsigned char *out, unsigned char *iv, uint8_t counter);
+    void generate_nonce(unsigned char *out, unsigned char *iv, uint64_t counter);
     void generate_write_key(unsigned char* out, unsigned char* secret);
     void generate_write_iv(unsigned char* out, unsigned char* secret);
     void get_hash(crypto_hash_sha256_state *ptr, unsigned char *out);
